Implement basic pretty printing
The code for pretty printing is in KidPrettyPrint.hs. What is there
should give a good model for how to extend it. KidParser will now print
the program as a data structure followed by the pretty printed version.
It is currently incomplete - inputting a program with unknown commands
will result in a "Non-exhaustive patterns" error. testfiles/print1.gt
will work with the current basic version.